# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  version: v1.0.0
  title: Attendance Application API
  description: API for Attendance application.
  x-audience: external-partner
servers:
  - url: 'http://{host}:{port}'
    description: http server
    variables:
      host:
        default: 'localhost'
        description: identity server host
      port:
        default: '8000'
        description: identity server port
  - url: 'https://{host}'
    description: https server
    variables:
      host:
        default: 'localhost'
        description: identity server domain
security:
  - AccessToken: []
...
paths:
...
  /api/v1/auth/login:
    post:
      summary: Authentication a user
      operationId: passwordLogin
      security: []
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: Successful login response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginApiResponse'
        "403":
          $ref: '#/components/responses/Forbidden'
        "400":
          $ref: '#/components/responses/BadRequest'
        "500":
          $ref: '#/components/responses/UnexpectedError'
  /api/v1/auth/logout:
    post:
      summary: Record user login and logout times
      operationId: userLogout
      tags:
        - Auth
      responses:
        "200":
          description: Successful logged time.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
        "403":
          $ref: '#/components/responses/Forbidden'
        "400":
          $ref: '#/components/responses/BadRequest'
        "500":
          $ref: '#/components/responses/UnexpectedError'
  /api/v1/auth/refresh-token:
    post:
      summary: Refresh token
      operationId: refreshToken
      security:
        - RefreshToken: []
      tags:
        - Auth
      responses:
        "200":
          description: Successfully refreshed jwt token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginApiResponse'
        "403":
          $ref: '#/components/responses/Forbidden'
        "500":
          $ref: '#/components/responses/UnexpectedError'
...
components:
  securitySchemes:
    AccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: The authorization token used by the user to authorize API requests.
    RefreshToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: The token used to refresh an expired access token.
    BasicAuth:
      type: http
      scheme: basic
      description: Basic authentication using username and password.
...
  schemas:
...
    ApiError:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: The server time at which this response was created.
        code:
          type: integer
          format: int32
          description: The status code the server returned.
        status:
          type: boolean
          description: Whether or not the request was a success or not.
          default: false
        message:
          type: string
          description: A short error message from server. If there are multiple errors they will be comma seperated
...
    GenericResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: The server time at which this response was created.
        code:
          type: integer
          format: int32
          description: The status code the server returned.
          example: 200
        status:
          type: boolean
          description: Whether or not the request was a success or not.
          default: true
        message:
          type: string
          description: A short success message from server
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: User account username.
          x-email-message: "{messages.usernameRequired}"
        password:
          type: string
          description: User account password.
          x-required-message: "{messages.passwordIsRequired}"
    LoginApiResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: The server time at which this response was created.
        code:
          type: integer
          format: int32
          description: The status code the server returned.
          example: 200
        status:
          type: boolean
          description: Whether or not the request was a success or not.
          default: true
        message:
          type: string
          description: A short success message from server
        data:
          type: array
          items:
            $ref: '#/components/schemas/TokenResponse'
...
    TokenResponse:
      type: object
      required:
        - token
      properties:
        accessToken:
          type: string
          description: user access token.
        refreshToken:
          type: string
          description: user refresh token.
...
  responses:
    BadRequest:
      description: Request syntax is invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            timestamp: 2023-02-21T18:53:37.709Z
            code: 400
            status: false
            message: "bad request"
...
    UnexpectedError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            timestamp: 2023-02-21T18:53:37.709Z
            code: 500
            status: false
            message: "something unexpected occured"
    Forbidden:
      description: Forbidden error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            timestamp: 2023-02-21T18:53:37.709Z
            code: 403
            status: false
            message: "forbidden"